Rewrite functions as proc-macros
The initial implementation of this crate used derive-macros. These were
easy to write, but made it very difficult to instrument with features.
For example, deriving serde's traits when a hypothetical `serde` feature
is enabled was not possible to implement. Or, at least, I did not know
how to do this ergonomically.

Switching to proc-macros increases the complexity and compile time of
this crate, but it also gives us much more tools to control the macros'
outputs. We expect that this will make it easier to implement complex
types.