Create macro for fields with secrets
A new macro has been created that generates a new type for fields that
contain secrest such as passwords or API tokens. The field is backed by
the `SecretString` type from the crate `secrecy`[^1], which prevents
the contents of the string to be printed in logs or debug statements.

Since this feature requires a new dependency, it must be enabled with a
feature flag.

[^1]: https://crates.io/crates/secrecy