Combine continuous integration into a single job
Now that we are using Earthly to configure our continuous integration, we want
to use it to optimize its execution as well. Earthly calculates a directed
acyclic graph with all steps, and then parallelizes them as much as possible.
Together with its aggressive caching, we expect this to cut down the execution
time of CI.