Ignore `non_local_definitions` warning
Recent versions of Rust warn that the code violates the non-local definitions
rule[^1]. The error message points to `serde::Deserialize`, which is outside our
control. It is not unlikely that our proc marcos are to blame, but that requires
further investigation.

For now, the warnings have been turned off and a `TODO` has been added to every
`allow` statement.

[^1]: https://doc.rust-lang.org/stable/nightly-rustc/rustc_lint/non_local_def/static.NON_LOCAL_DEFINITIONS.html