Unlock dependencies
This crate does not require specific versions for its dependencies. In
fact, in an effort to increase compatibility with its dependants, it is
tested against both the lowest and highest versions of its dependencies.

While this was always the intention, the `Cargo.toml` was not using the
correct way to specify the dependencies to actually support this
behavior. This has been fixed.