Run CI on GitHub Actions runners
The performance of Earthly's runners has been a serious bottleneck for
our continuous integration, to the point where builds have either timed
our or taken a few times longer than before.

While caching does not work as well on GitHub Actions runners, and we
have not configured it yet, we are testing the performance of those
runners to see if it improves the build times significantly enough to
switch back.