Include README in releases (#58)
When publishing a release, we want to include the README in the crate so
that it gets rendered on crates.io. But so far, we only copy Rust files
into the build container. A new step has been added to the publish
target to copy additional files into the container before publishing the
crate to crates.io.