Refactor Earthfile and GitHub Actions workflow
Earthly has been pushed to the absolute limits with this commit. The
top-level `Earthfile` has been broken up into smaller files by
reimplementing its targets as functions in different files inside the
`.earthly` directory. Because Earthly uses the parent directory of each
file as the context for each of its targets, the nested files can only
define functions that are executed with the context of the calling
target in the project root.

The root-level targets have been renamed to follow a simpler naming
convention, which has made it possible to dynamically discover which
targets to execute on GitHub Actions. Any target starting with `check`,
`format`, `lint`, or `test` will get executed in CI.

The formatting targets have also been refactored to optionally write
changes back to the project, making it possible to use them as pre-commit
hooks.