Restrict maximum dependency versions
In conversation with @martijnarts and @benwis, we discussed how to both provide
a wide range of dependency versions and still ensuring that breaking API are
caught early. The best solution seems to be to define a lower bound and restrict
the upper bound to the latest major version that has been tested against this
crate. When future major versions of dependencies are released, typed-fields can
be tested against them to ensure the crate supports them cleanly.