Replace Earthly with Flox + Just (#121)
Earthly has been deprecated[^1] a few months ago and will see no further updates
or support. While a community fork is trying to get off the ground, we are
looking for alternatives that are more actively supported. One approach that we
want to try is combining [Just] as a task runner with [Flox] environments. The
upside of this approach is that Flox offers a few more features for local
development, while the downside is the lack of task isolation and caching.

[flox]: https://flox.dev
[just]: https://just.systems/
[^1]: https://earthly.dev/blog/shutting-down-earthfiles-cloud/